+++ flask_hmac/exceptions.py
@@ -1,0 +1,15 @@
+
+class HmacException(Exception):
+    pass
+
+
+class SecretKeyIsNotSet(HmacException):
+    pass
+
+
+class InvalidSignature(HmacException):
+    pass
+
+
+class UnknownKeyName(HmacException):
+    pass