+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Change Log
 ----------
 
+0.1.1
+~~~~~~~~~
+- Able to change secret key in `make_hmac` method
+- Method `validate_signature` created which can be used outside of `Hmac` class
+- Custom exceptions
+
 0.0.1
 ~~~~~~~~~
 - Initial release including the core feature set